Set up a system with a Makefile, buildout.cfg, and tox.ini
Do `make toxtest` to run tox tests on Python 2.6 and Python 2.7
Do `make install` to do a `python setup.py install`.
Both of the above use buildout to download and build FreeTDS and the
Makefile sets environment variables that make sure that _mssql.so links
with the FreeTDS libraries that are built by buildout.